% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sexy1_coverage}
\alias{sexy1_coverage}
\title{Drone data from Tim, vegetation coverage}
\format{
A data frame with an expanding number of rows and 8 variables:
\describe{
\item{date}{the ymd the data was collected}
\item{plot_idtim}{the obs unit in Tim's notation, linked to sexy1_eu}
\item{mean}{I don't know the units on these...}
}
}
\usage{
sexy1_coverage
}
\description{
A dataset containing the data provided by Tim and his drone
regarding vegetation coverage. The variables are as follows:
}
\keyword{datasets}
