% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sexy1_standcounts}
\alias{sexy1_standcounts}
\title{Manually collected standcount data}
\format{
A data frame with an expanding number of rows and 4 variables:
\describe{
\item{trial_id}{Identifies which trial, Trial1 was established fall 2024}
\item{date}{the ymd the data was collected}
\item{eu2_id}{the obs unit the data was collected on}
\item{plants_m2}{The number of plants present per square meter}
}
}
\usage{
sexy1_standcounts
}
\description{
A dataset containing the data collected by Anders and Gina
estimating stand counts. The variables are as follows:
}
\keyword{datasets}
