% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEXYrye-package.R
\docType{package}
\name{SEXYrye-package}
\alias{SEXYrye}
\alias{SEXYrye-package}
\title{SEXYrye: SEXYrye Is A Data Package From The SEXY Trial}
\description{
The package doesn't do anything except make your data analysis life 10,000x easier.
}
\author{
\strong{Maintainer}: Gina Nichols \email{virginia.nichols@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
