% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEXYrye-package.R
\docType{package}
\name{SEXYrye-package}
\alias{SEXYrye}
\alias{SEXYrye-package}
\title{SEXYrye: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
